Extract methods about OpenAsset and move them to src/open_assets.rs
