+++ src/bin/electrs.rs
@@ -18,7 +18,8 @@ use electrs_tapyrus::{
     errors::*,
     index::Index,
     metrics::Metrics,
-    query::{OpenAssetCache, Query},
+    open_assets::OpenAssetCache,
+    query::Query,
     rpc::RPC,
     signal::Waiter,
     store::{full_compaction, is_fully_compacted, DBStore},
